v1.0.6: TG/Emby 通知对齐参考项目 + 渲染层统一
- Telegram: FormatMessage 业务优先级排序 + 批量合并冒号统一 + /scan/cleanup 跳转菜单 + /help 重写
- Emby: 播放停止 emoji ⏹️ + 播放进度字段名 + 播放通知仅 Primary 配图 + 详情请求条件化
- STRM/任务通知统一走 FormatMessage 三段式渲染
- 测试: 新增 27 个用例覆盖 FormatMessage/STRM 通知/Emby 模板/播放条件字段
- 文档: 新增 .changes/v1.0.6.md，更新 CHANGELOG、wiki 版本日志/TG通知/Emby集成
- 工具: sync_release_notes.py 移除硬编码 PAT（改读环境变量 GITHUB_TOKEN）

+++ internal/service/notify/menu.go
@@ -441,7 +441,7 @@ func (h *CommandHandler) RouteMenuCallback(ctx context.Context, cq CallbackQuery
 		return nil
 	}
 	chatID := fmt.Sprintf("%d", cq.Message.Chat.ID)
-	messageID := cq.Message.MessageID
+	messageID := int64(cq.Message.MessageID)
 	data := cq.Data
 
 	logger.S().Debugf("[Menu] Route callback: %s", data)
